+++ src/kernels.h
@@ -8,7 +8,7 @@ namespace ffi = xla::ffi;
 
 ffi::Error FooFwdHost(cudaStream_t stream, ffi::Buffer<ffi::F32> a,
                       ffi::Buffer<ffi::F32> b, ffi::ResultBuffer<ffi::F32> result,
-                      ffi::ResultBuffer<ffi::F32> b_plus_1, size_t n);
+                      ffi::ResultBuffer<ffi::F32> b_plus_1, int64_t n);
 
 ffi::Error FooBwdHost(cudaStream_t stream,
                       ffi::Buffer<ffi::F32> scalar_grad,
@@ -16,6 +16,6 @@ ffi::Error FooBwdHost(cudaStream_t stream,
                       ffi::Buffer<ffi::F32> b_plus_1,
                       ffi::ResultBuffer<ffi::F32> a_grad,
                       ffi::ResultBuffer<ffi::F32> b_grad,
-                      size_t n);
+                      int64_t n);
 
 #endif // KERNELS_H_
